[zbi_test] Set timeout for the host test to 10 min.
Some zbi tests set a timeout of 30s and the success string shows up just
around that timeout, but there's a little delay in seriallistener
reading the success string. Thus, testrunner kills the test just as
seriallistener finds the success string and so it looks like the test
passed but testrunner marks it as timed out.

We will revert the timeout to 10 minutes which was the timeout when the
tests were running on the zbi_tests builders before we ran them as host
tests. Once we have more data, we can determine an appropriate timeout
for the tests.

Bug: 97267
Change-Id: Ia688aa8833996399c9f93bad0d2ca58a29f98600
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669506
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>