[debugger] Ignore unsupported type signatures
When `-fdebug-types-section` is enabled, type definitions are put in a
separate section ".debug_types" and only a DW_AT_signature is left at
the original place. In DWARF5 .debug_types is merged into .debug_info
section.

We don't support it yet and logic to ignore them (and TODOs) are added.

Bug: 97388
Change-Id: I70e237830ea7203360dbb4f075cd722d7c32d888
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666549
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>