[phys] TrampolineBoot arm64 support.
Update build rules and ASM to build and test trampoline boot
for arm64 as well.

In arm the boot shim provides a ZBI_TYPE_DEVICETREE, which
is expected, but unnecessary for trampoline boot tests.

Because we are performing memory copies with caches turned off,
and the device tree item size is 1 MB in qemu, this causes huge
performance drops (x86 ~100 iters in 40 secs, while in arm64
we barely get ~20). This is due to a zbi size delta of ~1MB
between arm and x86 (devicetree).

By removing it on the bootstrapping step, we can reduce
the time per iteration, and avoid carrying around an
unnecessary item.

Bug: 88583 #done
Test: fx run-zbi-test trampoline-boot-zbi-test
Change-Id: I5b49219e51a531510405c6a1f4d160da44a40472
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659268
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>