[swd-e2e-tests] Pin workstation to a specific build
In fxr/661742, the storage team landed a CL that results in minfs being
rebuilt. This would cause the system to lose the authorized_keys file.
As an attempt to avoid this, this patch pins to a specific build, since
the last green build was from before fxb/661742 landed.

Once we have a green build, this CL can be reverted.

Bug: 95993
Change-Id: Id079e61ba63f2986daa1db9365ac513bd528f56d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670463
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>