[settings] Add proxy targets and reduce visibility
In preparation for migrating away from //garnet, this CL reduces the
default visibility of all targets in //garnet/bin/setui to only be
visible in the file itself.

Targets that are used externally have proxy targets added that scope
down the visibility to the minimum that's needed.

Fixed: 98204
Test: N/A, just modifying build visibility
Change-Id: I301f40a21bf99c9e89121403c54fd0042fac11aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669488
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: William Xiao <wxyz@google.com>