[kernel][vm] AwaitingClean pages beyond supply_zero_offset
Committed Dirty pages beyond supply_zero_offset are transitioned to
AwaitingClean on WritebackBegin. This ensures that a future write is
correctly intercepted to generate a dirty request.

These pages remain in AwaitingClean for now; they are not transitioned
to Clean on WritebackEnd. That transition will be added only when
supply_zero_offset can be advanced on WritebackEnd (done in a separate
commit).

Test: core-pager-writeback
Bug: 94371
Change-Id: I02d6f04dac59edc8b2f02bbe4c03ae8ff01eeda7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670009
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>