[fint] Upload build dir manifest to GCS on no-op failure
If the build doesn't converge to a no-op then it's useful to be able to
look at a manifest of the contents of the build directory, including
symlink destinations and mtimes of all files.

The file will contain a JSON array with elements of the form:

  {
    "path": "host_x64/foo",
    "mtime": 1648758742802.102,
    "symlink_target": "../../prebuilt/foo/foo"
  }

"symlink_target" will be omitted if the file is not a symlink.

Bug: 94334
Change-Id: I6d9326cdd150464e8e971d47264f044db8be13cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664297
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>