[ffx] Fix sdk-type error for emulator
When adjusting some error reporting code in the emulator startup code,
we overlooked that the in the SDK, a config value is omitted rather than
populated with a different value. This led the launcher to erroneously
fail when launched from the SDK.

This CL removes the invalid check and replaces it with the ffx config
API call which does all the appropriate error handling for us.

Change-Id: I6a2cc310d386c6d79a72ac00db11740777410ac1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670250
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>