[ffx] Tweak --log helptext
The helptext for the --log parameter is inaccurate, claiming that by
default all output goes to stdout/stderr. In truth, we set up a log file
in the working directory by default, and print that path to stdout. This
CL adjusts that helptext to indicate this.

Change-Id: I5410dad2bfaadc999d47d5c50d4449e9b983dd87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670583
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>