[drivers] Fix log severity configuration for drivers
This patch fixes a bug where anything in a driver below INFO
is not logged. This was caused by an inconsistency between
the new backend and the old backend. This patch adds a new
pass-through method used by the old backend which ignores the
log severity in the new backend.

Fixed: 98252

Test: fx test "//src/diagnostics/validator/logs/ddk:ddk-log-tests-package_test_ddk-log-tests-package_component(//build/toolchain/fuchsia:x64)"

Change-Id: I3724a34262ef364cbab0aaf3c13d41538a86780e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670063
Reviewed-by: Yilong Li <liyl@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>