[ffx][repo] Starting server should not check server mode
Back before we had `ffx repository server start`, we'd always start the
server in the background if we had any registered repositories. One of
the confusing errors we could get from that though is that if a user had
`pm serve` running, the repository server would silently fail to start.
The only way to tell this happened was to read the ffx logs.

Since then, we've added explicit controls to starting and stopping the
server, and migrated `fx` and `fserve` over to always use these
commands, rather than assume the ffx repository server is running in the
background. This allows us to report that the server couldn't start, and
help guide the user into how to rectify the issue.

Because of this, we no longer need to have ffx repository block running
the server based off the state of `repository.server.mode`. This patch
removes those checks, which should make it easier for users to try out
ffx repository server.

Fixed: 98349
Change-Id: I711536a156fb284085b7efacb8a58d1bff6e48ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669901
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>