[ffx] Change Result<i32> to Result<()>.
The i32 value in the Ok(_) for this result was unused in main.rs and in
the ffx_plugin_impl functions. The i32 values being set are not actually
read later on to determine the exit code (as can be seen in
https://cs.opensource.google/fuchsia/fuchsia/+/main:src/developer/ffx/lib/errors/src/lib.rs;l=130;drc=6f5b3cc7681a97819f7e58b01df63ef876161a48).

Given these values serve no purpose they have been removed to prevent
confusion to readers.

Bug: 98220

Change-Id: I36414841cedd4b8e61a08deb5590670c7fd918d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669835
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>