[media/vnext] allow Threads to be invalid
This change updates the fmlib::Thread class to allow a Thread to be
invalid/uninitialized. This allows a Thread to be used in a consistent
way even when it cannot be initialized immediately or when associated
resources must be released before the Thread goes out of scope.

Test: trivial change
Change-Id: I7ac557bf68daf2c9b128eb317ded525501e360b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670740
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>