[graphics][compute][target_archive] Disambiguate target_archive tool name on host
The $(host_toolchain) was conflicting with
`//build/toolchain:linux_64` because the tool names produced by the
target archive script shared the same triple { "cpu", "name", "os" }.

This workaround is to prefix the tool name with "host" when using the
host toolchain.

Testability:

The CL fxrev.dev/658743 revealed this issue and it was failing on CQ.

This CL fixes the issue.

Additionally, inspection of `<out>/tool_paths.json` shows that the
host target archive tool names are now prefixed with "host".

Change-Id: I054b22eac593ea2c1c683de4d093a3ac69651fa5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670944
Reviewed-by: Josh Gargus <jjosh@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>