[ermine] Remove dialoginfo after dialog is dismissed.
This change removes a dialoginfo from the list of dialogs after the
corresponding dialog was dismissed. It also fixes an issue when a stuck
app that failed to load (https://fxbug.dev/98083) would leave ermine in
an unresponsive state. This was because the `appIsLaunching` flag was
not reset.

Test: Tested manually
Bug: 98083
Change-Id: I11a1bca59b9eb057f4b3f7f0b6b9b7ef54a09ee6
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/671143
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Yeonhee Lee <yhlee@google.com>
Reviewed-by: Charles Whitten <cwhitten@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>