[kernel][socket] Change read off and buf length
This changes how read offsets and lengths are stored in each MBuf.
Previously there was both an offset and length remaining stored in
the MBuf, with the length being the amount of data, from that offset,
remaining. This meant that a read would both increment offset, and
decrement length.

This makes it so that the length is just the length of data in the
MBuf, and there is a single read offset stored in the MBufChain. A
single read offset is sufficient since only one MBuf (the current
buffers_.front()) is actually being read from. To then simplify the
MBuf invariants the length is changed from length remaining from the
offset, to just the length of data in the MBuf.

Change-Id: Ifcc27a624d3beedef7f1d4e343d687ec99415192
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670883
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>