[python][serialization] Add proper support for default values
When deserializing a dict into an instance of a class, if a field of the
class is not present in the dict, properly detect if it has a default
value, and allow that value to be used instead.

Found while converting a class to use serialization module.

Bug: 98339
Change-Id: I864bd8c959a039d2368b5226c6d4dac0526964e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669895
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>