[fidl] Use a common unknown_interactions.test.fidl
Move the protocols used in unknown interactions tests to a common fidl
file to be shared between language bindings. This helps ensure changes
for one language will also be reflected for other languages, and also
make it easier to write these tests since the protocol ordinals will
be the same regardless of the language.

Bug: 88366
Test: fx test fidl_rust_external_tests
Test: fx test fuchsia-pkg://fuchsia.com/dart-bindings-test#meta/dart-bindings-test.cm
Change-Id: I3b4c68d75269e6c7866589a04553473165cb21aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670960
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>