[python][serialization] support use of base classes
Before this, serializing a class with a base-class would only
cause the subclass's fields to be serialized, the base-class's
fields would be omitted.

This was due to typing.get_type_hints() only return the hints
for the class of an instance, not the hints for any of its
base classes.

This adds a DFS through the base classes of a type to build up
the full set of field types.

Bug: 98356
Change-Id: If86f28ffad7f010a09e792f55cca54ab2da7c9b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670447
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>