[fx-set] Write relative path to .fx-build-dir
fx seems to support this. Several fx tests rely on it, and the path is
turned into an absolute path inside vars.sh, so it shouldn't break
anything. Using a relative path in this file means that the Fuchsia
checkout can be moved without causing problems.

Change-Id: I9528f4eca271493ccc4b075274be8e1695f54579
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671241
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>