[sdk][stdcompat] Disable tests that take the address of STL functions
Changes in upstream Clang have enabled -Wc++20-compat by default, and
part of those checks is to now warn when a function from the standard
library has its address taken, but is unavailable. This occurs here
because `std::address_of` is a compiler builtin. In fact taking its
address has for some time resulted in clang generating a compatible
function definition for the address taken behavior to remain valid,
since it was acceptable in earlier standard versions.

The tests we are disabling are comparing function pointers for equality.
There are 2 basic issues here that make me suspicious of the test:

1. Pointer provenance can be surprising. AFAICT it is legal for the
   compiler to assume two function pointers are distinct if they
   originate from different names (even if they really do alias). Since
   that should be legal for the compiler to do, I’m not sure the test is
   truly reliable, though I’m unaware of exactly when that happens in
   practice.

2. Because the test seems to pull in the std::address_of definition in a
   single place, and that should be available in the same TU as the test
   code, I’m not sure it's possible for the test to ever fail, unless
   the behavior above is being enforced. Since I think all the code is
   available to be inlined, I would be surprised if the check wasn't
   largely optimized away since I expect that CSE would find the
   redundant load, and simply reuse it.

In the future we plan to address this using a bit of C++ cleverness.
Essentially we will ensure that compilation fails when the polyfill
version are used when a viable function could have been supplied from
the STL.

Since the polyfill versions are templates, we specialize them using a
private type, but mark those implementations as delete. Any attempt in
our tests to use the polyfill version, will fail at compile time, since
it is an error to make a call to a deleted function.

However, until that approach is ready, we simply disable these tests to
avoid build failures.

Bug: 98345
Change-Id: Iae6538bdcbdc5238406bf2ae9f3389cc51b4ad33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671787
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Paul Kirth <paulkirth@google.com>