[ffx][triage] Make triage plugin work OOT
This cl modifies how default configs are loaded when no config paths are
passed as arguments.

We traverse the parent dirs, till a fuchsia checkout is encountered
(.jiri_root/ exists). The default configs in that case are
$FUCHSIA_DIR/src/diagnostics/config/... .

If no fuchsia checkout is found in the parent directories, the default
configs are expected to be in triage.config_paths variable which can be
set by `ffx config set triage.config_paths "[\"default.triage\",
\"default2.triage\"]"`

BUG: 95665
Change-Id: I1bf2972c4fdaf472f2c3f2cc56774667114efc1a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669742
Reviewed-by: Chris Phoenix <cphoenix@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Pawan Dogra <padogra@google.com>