[ffx] Daemon reloads all manual targets on restart
The attached bug describes a workflow where an emulator is dropped from
the ffx daemon's target collection when the daemon restarts, even though
it's still running. I tracked this down to an extra timeout check in the
reload code.

This CL removes the timeout check, and allows the targets to be loaded
regardless of timeout. The daemon's target expiry loop will then remove
and clean up any targets that are no longer responsive.

Bug: fxbug.dev/98510
Change-Id: I81cec8b4107a88bbb83cf8f6319d7b7bec7b859a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671867
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>