[graphics][compute] Update spinel-vk-rs Rust bindings to point at Spinel2
The `spinel-vk-rs` Rust bindings now point to Spinel2.

Additionally, because in-tree Rust currently has no awareness of
Vulkan 1.2, a strawman small C wrapper has been created to hide the
1.2 dependencies.

The Spinel backend has been temporarily removed from Carnelian. This
wrapper will be maintained and improved as changes are landed to
Spinel so that reintegration will be simpler than before.

Note that this is a work-in-progress "helper" library and and when
it's integrated with Carnelian it will be clear which Vulkan state
must be made public.

Testability:

* fx shell run spinel-vk-rs-tests

Change-Id: I77c65062c95ef306bf74ce72af0ab4771ad7f0b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658743
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Dragoș Tiselice <dtiselice@google.com>
Commit-Queue: Allan MacKinnon <allanmac@google.com>