[dgramsocket_test] Test clearing errors with cache
Fast UDP uses a single `zx_object_wait_many` to check for 1) errors
asserted on the zx socket and 2) invalidation of {destination,cmsg}
cache elements. Add tests that verify that I/O methods return and clear
an error whenever it is present on the socket, regardless of the state
of the Fast UDP caches.

Fixed: 98073

Change-Id: Id2a5e1448d26ceaa4b15648c67c6acf023dd00e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/668960
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Arthur Sfez <asfez@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>