[codec][vaapi] Fixed VP9 decoder issue
The VP9 decoder was failing a unit test where the first frame of the
stream was not a intra frame and instead was an inter frame. The decoder
was trying to reference frames that did not exist. Now the decoder will
keep parsing stream data until it encounters an intra frame. Then inter
frames will now have valid references.

Fixed: 98527

Change-Id: I1108efbb8a8b2a388265934139b97b6d329c78a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671892
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>