[test-manager] Close event senders as soon as possible
The latest hang in a coverage builder shows that ffx test is waiting
for more run events from test_manager, but all coverage has already
been drained. test_manager knows that there are no more events to send
to the client when an internal mpsc channel is closed.
This is a speculative fix that closes the channels as early as possible
to minimize the chance of run events hanging.

Bug: 94333

Change-Id: Ie4aa996a2324d7635dd16cbc63adad76eab140f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671786
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>