[benchmarks] Add checks to enforce a naming scheme for performance metrics
Enforce some constraints on the metric names that can be used in the
"test_suite" and "label" fields in fuchsiaperf.json files.

While this could be implemented in catapult_converter (in C++), this
is easier to implement (and easier to write tests for) in Dart.

Change the existing tests for convertResults() to create real
fuchsiaperf files rather than using filenames for files that don't
exist.  This is necessary now that convertResults() reads the files
itself.

Bug: none
Change-Id: I775754051621a836c692a85450e3bcc86114a677
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672088
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>