[ffx][pbms] Workaround terminal product naming issues
This addresses the issue in fxbug.dev/98529 which reports
issues when downloading the packages for the terminal.qemu-arm64
product.

This handles 2 aspects, first mapping the name "terminal" to the legacy
tarball name of "qemu". Second, pulls the version from the product
bundle metadata vs. from the SDK.

Bug: 98529

Change-Id: I099f7274614aed90174d38280eda502dddbfeb84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672124
Reviewed-by: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>