[ffx] Add kernel_args flag to ffx emu
The attached bug indicates a high-priority need for the ability to
specify additional arguments when launching an emulator. This was
previously handled via the "--kernel_args" flag in fvdl. We have a
project planned to make configuration easier (see fxbug.dev/98268), but
it won't be complete in time for the training sessions described in the
attached bug.

This CL adds a command line flag to support the old behavior of adding
extra args on the command line that are passed through to the emulator
without additional processing. The flag is marked "experimental" in the
helptext to indicate to end-users that this flag is not intended to be
used in permanent workflows such as infrastructure tests. We intend to
remove this flag when fxbug.dev/98268 is complete.

Bug: fxbug.dev/95278
Change-Id: Ifd06cc020f275ac994c44754f8bddab5937bd37e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672746
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>