[identity] Verify root hash in the tree is the same as loaded value.
This change calls get_log and verifies the hash in the cr50 is the
same that in the merkle tree. If it isn't the same it fails and fails
out from launching the credential_manager. This is strictly better
than having a sync and having every call fail to the credential
manager.

Bug: 95277
Change-Id: I3c0af89606b5c49da1151829e3c58b9d3fb61f52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670008
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Ben Wright <benwright@google.com>