[starnix][signals] Add support for restarting syscalls
When a signal interrupts a syscall, and the signal action handling the
signal has the flag `SA_RESTART`, the syscall should be restarted if the
syscall supports it.

This change adds the machinery for restarting syscalls, and implements
it for the `read`, `write`, `wait4`, and `waitid` syscalls.

Multiply: test_chromiumos_distro

Change-Id: Ia6833a1cd5c3dc622bf6bc6ad25b0ec186f76526
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671326
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>