[testrunner] Run snapshot even if we hit a timeout.
The detached bug describes an issue where some tasks time out because
ffx fails to respect the test timeout and hangs. In order to debug, it'd
be helpful to have the snapshot available even if the context was
cancelled due to a timeout.

Bug: 94333
Change-Id: Ic937508b707e1182eb3f2bca2676bddc3caf89ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673090
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>