[benchmarks] Convert dart-fidl-benchmarks to be run via Components v2
Use the Dart test wrapper in src/dart/testing/ for this.

Change that wrapper to pass through the custom_artifacts capability,
because that is used for getting performance results out of the test.

Update the README.md by removing instructions for running the test
that are no longer applicable, because they pre-date the plumbing in
fidl_microbenchmarks_test.dart for reporting the results through the
performance testing infrastructure, and I don't think it makes sense
for each test to have instructions on how to run it.  These
instructions were originally in the Topaz repo under
tests/dart_fidl_benchmarks/README.md.

Bug: 98667
Change-Id: I3e680cc6581b2ddc238065c3139dfcc463b4fb59
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672283
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>