[audio][drivers][usb-audio] Fix frame rate range handling
If a USB audio device specifies a range of frame rates supported
as in some versions of the Samson Go Mic, then we need to extract
valid rates to use. This CL fixes the handling of this continuous
range by generating valid rates (in either the 44.1Khz or 48kHz
family) from the range specified.

Fixed: 98153

Test: usb-audio-test

Change-Id: I18c4f7da7009a8dfdacf24ce827a2fff7ff3202c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672003
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>