[CredMgr] Add diagnostics for hash tree operations.
This adds the diagnostics operations for things that are generally
fallible, i.e. loading and saving (update/delete operations shouldn't
fail unless we've got a significant bug in the module which we'd find
through other means). It also adds a credential-count event which
tracks the number of credentials in the hash tree.

Bug: 91714
Change-Id: I271e089808c724537e0edf9a6ae3385c399db63e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672090
Reviewed-by: Daniel Hwang <arkay@google.com>
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>