[CredMgr] Add diagnostics for the pinweaver RPCs.
I think this is the last main block of functionality we need so I'm
planning to close the bug after this one. We should probably add a
couple more bits of information as we wrap up the log replay but that
needs a little bit more thinking to work out where they best fit.

Bug: 91714
Change-Id: I61d1710a7206a07841485d62127ae67da70d5f78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/671945
Reviewed-by: Daniel Hwang <arkay@google.com>
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>