[fidl] Move test types into sdk
It is not currently possible to use the common
unknown_interactions.test.fidl from HLCPP because it is defined in /src
rather than /sdk. Moving these test files to /sdk allows them to be used
in sdk tests, which currently includes HLCPP, but may also need to
include LLCPP Wire and Natural Types as well as Rust in the relatively
near future.

Bug: 88366
Change-Id: I7834b434de40a31696d9b643cd444238c1c9d69c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673067
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>