[firmware] add cipd_utils.py script
This script will provide some common functionality to interact with CIPD
prebuilts - up until now we've had a bunch of scripts implement this
functionality separately in one form or another, it'll be better to keep
it in a central place with tests.

To start with, this just provides the "changelog" functionality, which
takes two versions of a CIPD package, along with the source tree, and
generates a changelist of all git repos suitable for use in a git commit
message.

Example usage:

  $ ./cipd_utils.py changelog \
      ~/path/to/repo \
      cipd/package/name \
      old_cipd_version \
      new_cipd_version

  -- Changelist --
  repo1:
  [repo1] commit 1
  [repo1] commit 2

  repo2:
  [repo2] commit 1
  [repo2] commit 2
  [repo2] commit 3

  -- Source Revisions --
  repo1: rev1_B
  repo2: rev2_B

Bug: 99001
Change-Id: Iaaf526d50774ba546151a2bae9070294e9fc72e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/672844
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>