[starnix] Avoid lock cycle in ProcMapsFile
Take the lock inside the closure in ProcMapsFile::read_at
to avoid a re-entrant lock when writing out the result.
Previously, the lock was taken at the start of read_at,
which meant that it was being held throughout the entirety
of the function, including when the results were written
out to the task's memory.

Change-Id: I877d1b373d0b8144c7b4a97f593f20afb9c38f89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673669
Reviewed-by: Theodore Dubois <tbodt@google.com>
Commit-Queue: Steve Austin <steveaustin@google.com>