[audio][test] Fix audio_core_hardware_tests flake
These tests wait for audio_core startup, and for it to
identify the default audio input device. The tests then
fail if any device arrival/departure occurs during testing.

However, this could happen while the built-in audio OUTPUT
device was still be in the process of being discovered,
initialized and added. This would not be a problem if the
audio_core service was already started, but in certain
test configurations where this test was the first to
reference audio_core, these tests would often fail.

Now we ignore OnDeviceAdded if the device is an output;
we ignore OnDeviceRemoved if it is not our default audio
capture device; we ignore OnDefaultDeviceChanged if we are
not changing away from our default audio capture device.

This CL also clarifies the warning text that we emit when
we detect that the audio hardware device has a detectable
DC bias (we use more direct language that the hardware
should be retired).

Multiply: audio_core_hardware_tests: 30

Fixed: 81544
Fixed: 90834

Change-Id: I55740be176f04e4addcb8e009018dec56133c9f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673802
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Fuchsia-Auto-Submit: Martin Puryear <mpuryear@google.com>