[starnix] Refactor directory renames
This moves the check on whether the target is an empty directory when the
source is a directory to FileSystem::rename.

This allows to make the check and the replacement atomic, without taking
a lock on the replaced DirEntry that creates a lock order inversion.

Change-Id: If23023ae3ea9f95cf1ba3ea75c6a66545f46a743
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673424
Commit-Queue: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>