[virtualization] replace 'blocking::Unblock' with fdio file conversion
This CL removes blocking::Unblock and used EventedFd's to ensure the
CLI tool is only using sockets to communicate with Guests. This enables
consistency in how we both test the tool, as well as how we handle
all three data streams that are used when communicating with a Guest.

Testing: fx test guest-tests
Bug: 89427
Change-Id: Ie45bdf6f9fb30927f2dbfbc9f486e97f14b7a04e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658023
Commit-Queue: Jenna Macdonald <polyrain@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>