[cipd_utils] add repo spec and local CIPD packages
Adds support for two new features:

1. Repo spec
A repo spec allows us to select and optionally rename a subset of git
repos rather than using the entire manifest. This is necessary for Atlas
where we have a ton of unused repos we don't care about tracking, and
where the names we're using don't currently match the manifest name.

2. Local CIPD packages
We sometimes want to diff an existing CIPD package against a local
package before uploading it, so we can verify the changelog. Extend the
tool to allow either fetching CIPD packages from the server or using a
local CIPD package.

Bug: 99001
Change-Id: I7163f4d1baca6a8d902652ac2449c0eb57a668cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/673702
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>