[input] Wires up the focus distribution to the input pipeline.
The focus distribution went unused for the longest of times, but should not
be unused. The reason this went unused was that Ermine session shell did its
own routing, which it should not be doing. We're now fixing the issue.

Since proper routing for focus distribution is now split across multiple
repositories, we need a soft transition to reroute the relevant FIDL protocols
from Ermine back to the input pipeline (scene_manager in this case).  This is
the first in a series of changes required to make this happen.

Multiply: experiences_ermine_terminal_e2e_test
Multiply: experiences_ermine_terminal_smoke_test
Fixed: 98412
Bug: 95901
Change-Id: Ifee0eee4d63fcf74a06124143fb939b6776b808e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/666844
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>